move controller serving to controller package
