+++ pkg/cmd/server/origin/master.go
@@ -1,8 +1,6 @@
 package origin
 
 import (
-	"crypto/tls"
-	"crypto/x509"
 	"io"
 	"net/http"
 	"net/url"
@@ -13,24 +11,15 @@ import (
 	"gopkg.in/natefinch/lumberjack.v2"
 
 	apiextensionsinformers "k8s.io/apiextensions-apiserver/pkg/client/informers/internalversion"
-	utilwait "k8s.io/apimachinery/pkg/util/wait"
 	apifilters "k8s.io/apiserver/pkg/endpoints/filters"
 	apirequest "k8s.io/apiserver/pkg/endpoints/request"
 	apiserver "k8s.io/apiserver/pkg/server"
 	apiserverfilters "k8s.io/apiserver/pkg/server/filters"
-	"k8s.io/apiserver/pkg/server/healthz"
-	genericmux "k8s.io/apiserver/pkg/server/mux"
-	genericroutes "k8s.io/apiserver/pkg/server/routes"
-	authzwebhook "k8s.io/apiserver/plugin/pkg/authorizer/webhook"
-	clientgoclientset "k8s.io/client-go/kubernetes"
 	aggregatorapiserver "k8s.io/kube-aggregator/pkg/apiserver"
 	kubeapiserver "k8s.io/kubernetes/pkg/master"
 	kcorestorage "k8s.io/kubernetes/pkg/registry/core/rest"
 
-	"github.com/openshift/origin/pkg/authorization/authorizer"
 	configapi "github.com/openshift/origin/pkg/cmd/server/api"
-	serverauthenticator "github.com/openshift/origin/pkg/cmd/server/authenticator"
-	"github.com/openshift/origin/pkg/cmd/server/crypto"
 	serverhandlers "github.com/openshift/origin/pkg/cmd/server/handlers"
 	kubernetes "github.com/openshift/origin/pkg/cmd/server/kubernetes/master"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
@@ -318,113 +307,6 @@ func (c *MasterConfig) buildHandlerChain(assetConfig *AssetConfig) (func(http.Ha
 
 		return handler
 	}, nil
-}
-
-// TODO refactor this out of this package and split apiserver and controllers for good!
-func RunControllerServer(servingInfo configapi.HTTPServingInfo, kubeExternal clientgoclientset.Interface) error {
-	clientCAs, err := getClientCertCAPool(servingInfo)
-	if err != nil {
-		return err
-	}
-
-	mux := genericmux.NewPathRecorderMux("master-healthz")
-
-	healthz.InstallHandler(mux, healthz.PingHealthz)
-	initReadinessCheckRoute(mux, "/healthz/ready", func() bool { return true })
-	genericroutes.Profiling{}.Install(mux)
-	genericroutes.MetricsWithReset{}.Install(mux)
-
-	// TODO: replace me with a service account for controller manager
-	tokenReview := kubeExternal.AuthenticationV1beta1().TokenReviews()
-	authn, err := serverauthenticator.NewRemoteAuthenticator(tokenReview, clientCAs, 5*time.Minute)
-	if err != nil {
-		return err
-	}
-	sarClient := kubeExternal.AuthorizationV1beta1().SubjectAccessReviews()
-	remoteAuthz, err := authzwebhook.NewFromInterface(sarClient, 5*time.Minute, 5*time.Minute)
-	if err != nil {
-		return err
-	}
-
-	// requestInfoFactory for controllers only needs to be able to handle non-API endpoints
-	requestInfoResolver := apiserver.NewRequestInfoResolver(&apiserver.Config{})
-	// the request context mapper for controllers is always separate
-	requestContextMapper := apirequest.NewRequestContextMapper()
-	authorizationAttributeBuilder := authorizer.NewAuthorizationAttributeBuilder(requestContextMapper, requestInfoResolver)
-
-	// we use direct bypass to allow readiness and health to work regardless of the master health
-	authz := serverhandlers.NewBypassAuthorizer(remoteAuthz, "/healthz", "/healthz/ready")
-	handler := serverhandlers.AuthorizationFilter(mux, authz, authorizationAttributeBuilder, requestContextMapper)
-	handler = serverhandlers.AuthenticationHandlerFilter(handler, authn, requestContextMapper)
-	handler = apiserverfilters.WithPanicRecovery(handler)
-	handler = apifilters.WithRequestInfo(handler, requestInfoResolver, requestContextMapper)
-	handler = apirequest.WithRequestContext(handler, requestContextMapper)
-
-	serveControllers(servingInfo, handler)
-	return nil
-}
-
-// serve starts serving the provided http.Handler using security settings derived from the MasterConfig
-func serveControllers(servingInfo configapi.HTTPServingInfo, handler http.Handler) {
-	timeout := servingInfo.RequestTimeoutSeconds
-	if timeout == -1 {
-		timeout = 0
-	}
-
-	server := &http.Server{
-		Addr:           servingInfo.BindAddress,
-		Handler:        handler,
-		ReadTimeout:    time.Duration(timeout) * time.Second,
-		WriteTimeout:   time.Duration(timeout) * time.Second,
-		MaxHeaderBytes: 1 << 20,
-	}
-
-	clientCAs, err := getClientCertCAPool(servingInfo)
-	if err != nil {
-		glog.Fatal(err)
-	}
-
-	go utilwait.Forever(func() {
-		glog.Infof("Started health checks at %s", servingInfo.BindAddress)
-
-		if configapi.UseTLS(servingInfo.ServingInfo) {
-			extraCerts, err := configapi.GetNamedCertificateMap(servingInfo.NamedCertificates)
-			if err != nil {
-				glog.Fatal(err)
-			}
-			server.TLSConfig = crypto.SecureTLSConfig(&tls.Config{
-				// Populate PeerCertificates in requests, but don't reject connections without certificates
-				// This allows certificates to be validated by authenticators, while still allowing other auth types
-				ClientAuth: tls.RequestClientCert,
-				ClientCAs:  clientCAs,
-				// Set SNI certificate func
-				GetCertificate: cmdutil.GetCertificateFunc(extraCerts),
-				MinVersion:     crypto.TLSVersionOrDie(servingInfo.MinTLSVersion),
-				CipherSuites:   crypto.CipherSuitesOrDie(servingInfo.CipherSuites),
-			})
-			glog.Fatal(cmdutil.ListenAndServeTLS(server, servingInfo.BindNetwork, servingInfo.ServerCert.CertFile, servingInfo.ServerCert.KeyFile))
-		} else {
-			glog.Fatal(server.ListenAndServe())
-		}
-	}, 0)
-}
-
-func getClientCertCAPool(servingInfo configapi.HTTPServingInfo) (*x509.CertPool, error) {
-	if !configapi.UseTLS(servingInfo.ServingInfo) {
-		return nil, nil
-	}
-
-	roots := x509.NewCertPool()
-	// Add CAs for API
-	certs, err := cmdutil.CertificatesFromFile(servingInfo.ClientCA)
-	if err != nil {
-		return nil, err
-	}
-	for _, root := range certs {
-		roots.AddCert(root)
-	}
-
-	return roots, nil
 }
 
 // InitializeObjects ensures objects in Kubernetes and etcd are properly populated.
